Move database query related types to IDataStorage
