+++ src/frontend/containers/AppToolbar/Menus.tsx
@@ -6,7 +6,7 @@ import { MenuButton, MenuRadioGroup, MenuRadioItem } from 'widgets/menus';
 import { getThumbnailSize } from '../ContentView/utils';
 import { MenuDivider, MenuSliderItem } from 'widgets/menus/menu-items';
 import { useStore } from 'src/frontend/contexts/StoreContext';
-import { FileOrder, OrderDirection } from 'src/api/FileDTO';
+import { FileOrder, OrderDirection } from 'src/api/IDataStorage';
 
 // Tooltip info
 const enum Tooltip {
